requirements.txt
# WhatsApp AI RAG Chatbot - Dependencies
# Core LLM & Embeddings
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10

# Vector Database
faiss-cpu>=1.7.4

# Document Processing
pypdf>=3.17.0
python-docx>=1.1.0
tiktoken>=0.5.0

# Web Interface
streamlit>=1.29.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.24.0

# Optional: For WhatsApp Integration (future)
# twilio>=8.10.0
